refactor: split config handlers by domain into 10 focused modules
